--- components/Form/InputText/index.tsx
@@ -1,3 +1,0 @@
-import InputText from './InputText';
-
-export default InputText;
