--- components/Form/InputTextarea/index.tsx
@@ -1,3 +1,0 @@
-import InputTextarea from './InputTextarea';
-
-export default InputTextarea;
