+++ components/Form/InputTextareas/index.tsx
@@ -1,0 +1,3 @@
+import InputTextareas from './InputTextareas';
+
+export default InputTextareas;
