+++ components/Form/InputTexts/index.tsx
@@ -1,0 +1,3 @@
+import InputTexts from './InputTexts';
+
+export default InputTexts;
